requirements/dev.txt
-r base.txt
factory_boy
pytest
pytest-django
pytest-flake8
pytest-cov
ipython
